Introduce subfolders, one aug per file
