+++ kornia/augmentation/_2d/__init__.py
@@ -1,27 +1,3 @@
-from .affine import RandomAffine
-from .box_blur import RandomBoxBlur
-from .center_crop import CenterCrop
-from .channel_shuffle import RandomChannelShuffle
-from .color_jitter import ColorJitter
-from .crop import RandomCrop
-from .cutmix import RandomCutMix
-from .earasing import RandomErasing
-from .elastic_transform import RandomElasticTransform
-from .equalize import RandomEqualize
-from .fisheye import RandomFisheye
-from .flip import RandomHorizontalFlip, RandomVerticalFlip
-from .gaussian_blur import RandomGaussianBlur
-from .gaussian_noise import RandomGaussianNoise
-from .grayscale import RandomGrayscale
-from .invert import RandomInvert
-from .mixup import RandomMixUp
-from .motion_blur import RandomMotionBlur
-from .normalize import Denormalize, Normalize
-from .pad import PadTo
-from .perspective import RandomPerspective
-from .posterize import RandomPosterize
-from .resized_crop import RandomResizedCrop
-from .rotation import RandomRotation
-from .sharpness import RandomSharpness
-from .solarize import RandomSolarize
-from .thin_plate_spline import RandomThinPlateSpline
+from kornia.augmentation._2d.geometric import *
+from kornia.augmentation._2d.intensity import *
+from kornia.augmentation._2d.mix import *
