+++ kornia/augmentation/_2d/geometric/__init__.py
@@ -1,0 +1,13 @@
+from kornia.augmentation._2d.geometric.affine import RandomAffine
+from kornia.augmentation._2d.geometric.center_crop import CenterCrop
+from kornia.augmentation._2d.geometric.crop import RandomCrop
+from kornia.augmentation._2d.geometric.elastic_transform import RandomElasticTransform
+from kornia.augmentation._2d.geometric.fisheye import RandomFisheye
+from kornia.augmentation._2d.geometric.horizontal_flip import RandomHorizontalFlip
+from kornia.augmentation._2d.geometric.pad import PadTo
+from kornia.augmentation._2d.geometric.perspective import RandomPerspective
+from kornia.augmentation._2d.geometric.posterize import RandomPosterize
+from kornia.augmentation._2d.geometric.resized_crop import RandomResizedCrop
+from kornia.augmentation._2d.geometric.rotation import RandomRotation
+from kornia.augmentation._2d.geometric.thin_plate_spline import RandomThinPlateSpline
+from kornia.augmentation._2d.geometric.vertical_flip import RandomVerticalFlip
