+++ kornia/augmentation/_2d/geometric/center_crop.py
