+++ kornia/augmentation/_2d/geometric/resized_crop.py
