+++ kornia/augmentation/_2d/intensity/__init__.py
@@ -1,0 +1,14 @@
+from kornia.augmentation._2d.intensity.box_blur import RandomBoxBlur
+from kornia.augmentation._2d.intensity.channel_shuffle import RandomChannelShuffle
+from kornia.augmentation._2d.intensity.color_jitter import ColorJitter
+from kornia.augmentation._2d.intensity.denormalize import Denormalize
+from kornia.augmentation._2d.intensity.earasing import RandomErasing
+from kornia.augmentation._2d.intensity.equalize import RandomEqualize
+from kornia.augmentation._2d.intensity.gaussian_blur import RandomGaussianBlur
+from kornia.augmentation._2d.intensity.gaussian_noise import RandomGaussianNoise
+from kornia.augmentation._2d.intensity.grayscale import RandomGrayscale
+from kornia.augmentation._2d.intensity.invert import RandomInvert
+from kornia.augmentation._2d.intensity.motion_blur import RandomMotionBlur
+from kornia.augmentation._2d.intensity.normalize import Normalize
+from kornia.augmentation._2d.intensity.sharpness import RandomSharpness
+from kornia.augmentation._2d.intensity.solarize import RandomSolarize
