+++ kornia/augmentation/_2d/intensity/earasing.py
