+++ kornia/augmentation/_2d/intensity/equalize.py
