+++ kornia/augmentation/_2d/intensity/gaussian_blur.py
