+++ kornia/augmentation/_2d/intensity/invert.py
