+++ kornia/augmentation/_2d/intensity/solarize.py
