+++ kornia/augmentation/_2d/mix/__init__.py
@@ -1,0 +1,2 @@
+from kornia.augmentation._2d.mix.cutmix import RandomCutMix
+from kornia.augmentation._2d.mix.mixup import RandomMixUp
