+++ kornia/augmentation/_3d/__init__.py
@@ -1,8 +1,2 @@
-from kornia.augmentation._3d.affine import RandomAffine3D
-from kornia.augmentation._3d.center_crop import CenterCrop3D
-from kornia.augmentation._3d.crop import RandomCrop3D
-from kornia.augmentation._3d.equalize import RandomEqualize3D
-from kornia.augmentation._3d.flip import RandomDepthicalFlip3D, RandomHorizontalFlip3D, RandomVerticalFlip3D
-from kornia.augmentation._3d.motion_blur import RandomMotionBlur3D
-from kornia.augmentation._3d.perspective import RandomPerspective3D
-from kornia.augmentation._3d.rotation import RandomRotation3D
+from kornia.augmentation._3d.geometric import *
+from kornia.augmentation._3d.intensity import *
