+++ kornia/augmentation/_3d/geometric/__init__.py
@@ -1,0 +1,8 @@
+from kornia.augmentation._3d.geometric.affine import RandomAffine3D
+from kornia.augmentation._3d.geometric.center_crop import CenterCrop3D
+from kornia.augmentation._3d.geometric.crop import RandomCrop3D
+from kornia.augmentation._3d.geometric.depthical_flip import RandomDepthicalFlip3D
+from kornia.augmentation._3d.geometric.horizontal_flip import RandomHorizontalFlip3D
+from kornia.augmentation._3d.geometric.perspective import RandomPerspective3D
+from kornia.augmentation._3d.geometric.rotation import RandomRotation3D
+from kornia.augmentation._3d.geometric.vertical_flip import RandomVerticalFlip3D
