+++ kornia/augmentation/_3d/geometric/center_crop.py
