+++ kornia/augmentation/_3d/intensity/__init__.py
@@ -1,0 +1,2 @@
+from kornia.augmentation._3d.intensity.equalize import RandomEqualize3D
+from kornia.augmentation._3d.intensity.motion_blur import RandomMotionBlur3D
