+++ kornia/augmentation/_3d/intensity/equalize.py
